osx 10.9 --debug update, part 1.
for osx 10.9 I haven't figured out how to perform the "run" operation
from lldb and have it execute as one would expect. Ironically, setting
up the whole show and falling off the end of the script allowing the
user to type run (or just r) actually works. So to allow for debug I
have temporarily moved to not starting the process in the script, but
allowing the user to do that.

It's unfortunate. I've tried many ways to "run" the process and
passing the binary on the command line etc. I tried passing the env
variable DISPLAY just in case that was it. There is something causing
FontForge to silently not start when the run is issued from the lldb
--script itself.