osx: improve debug output.
source files are now matched up to the binary offsets giving a clearer
picture to developers where an issue is happening.

I'm not overjoyed with how the paths are changed to the object files,
but finding a tool which specializes in doing this seems to have
evaded me so far. Some solution is better than no solution.