Update osx/FontForge.app/Contents/MacOS/FontForge
In https://github.com/fontforge/fontforge/issues/2055 I reported a dock bounce issue. This resolves this, by adding a `&` to the final command. The effect is that now when run the FF dock icon disappears after loading. This is helpful because the XQuartz icon is actually what users must click to bring FF windows to the foreground; and if they Force Quit XQuartz then FF will also exit. 

`/Applications/FontForge.app/Contents/MacOS/FontForge --debug` works with this change.