Fix crash in "MM" -> "Blend to New Font".
Selecting "Blend to New Font" in the "MM" menu currently crashes due to
detection of an invalid free. The issue is that MMChangeBlend calls
GCDFillupMacWeights which fills 'axisnames' with strings, but only the
first 'axis_count' should be freed (the rest are set to static
strings). The other caller of GCDFillupMacWeights (EditStyleName)
already does the right thing in this situation, which is to only free
the first 'axis_count' strings in 'axisnames'. This changes the cleanup
code in MMChangeBlend to do the same.