Fix crash caused by pressing escape in a charview without a wordlist
Pressing escape is meant to switch focus between the current tool and
the wordlist textbox. But this event can activate when the charview
is used without a wordlist textbox (e.g. the expand stroke dialogue).
This means that cv->charselector can be NULL, which leads to a segfault
when attempts are made to set the focus on it.

This fixes the 'esc' crash described in #2384 and #2029.