Fix temporary file creation on Windows.
Sigh... there is no mkstemp on Windows. The one that exists
in GLib can't be used because the file handle it returns
belongs to a different CRT instance, for some reason or
another.

Plus, you cannot unlink a file that is in use on Windows.

So roll our own by calling CreateFile directly. The flags
passed to it will ensure that the file created is only
readable by FontForge and that it will be deleted when
the file handle is closed.

This should give the intended behaviour for this function
on Windows.

Conflicts:
	fontforge/sfd.c