Allow the toolbar to use "selected" images when they are clicked.
This does currently require the theme to have the selected images in
it. A fallback to creating the selected images somehow when the theme
doesn't explicitly provide them still has to be done if we want to
maintain compatibility with old themes.

Note that the selected image has a bright red or bright green
background to be really obvious right now. In the future that will be
a more visually appealing selected image.