ofl.c and ofl.h created 2007-03-12
George probably didn't want to copyright name to the file, or forgot to
add it, but it's basically data files holding text information - which
can then be used to make a SIL open source license.

This patch resolves issue 2643 about author and file licence missing
when run through debian lint.