Moved is_LIGATURES into Unicode library instead of fontforge library
The "python" program to build these files contributed to a headache
worth of problems earlier between 2012 and 2014 when Fontforge built
python as an "optional" choice and not a "forced" dependency. Removing
the python program allowed us to continue onto 20140101 and onwards...
Example problems/issues #384, #387, #229, #565, #382, #241, #249, #233

makeutype.c already uses the Unicode list, so it can do this here too.