GDraw: Rework popup cleanup to avoid needing the call to GDrawProcessWindowEvents GDrawProcessWindowEvents is meant to process all window events pertaining to the given window. The call to GDrawDestroyWindow and GDrawSync beforehand *should* mean that the window in question is destroyed by the time that GDrawProcessWindowEvents finishes. However, this is not the case on GDK with the Win32 and Quartz backends, where GDrawSync basically does nothing.
Since waiting for an asynchronous event to occur is not particularly reliable
(or efficient on Win32/Quartz), instead rework the cleanup mechanism so that
et_close messages are sent to the popup in question. Since GDrawPostEvent
is implemented synchronously, we're guaranteed that the popup has run its
cleanup routines. Some minor modifications to the way in which the popup
handler works - by setting its user data to NULL after running through
the et_close cleanup procedure, means that when the et_destroy message is
received, even if it is received after its owner has been destroyed,
the handler can detect this, and handle the situation gracefully.