File chooser: Make the right click context menu work Fixes #2691.
If initiated from a right click, the GGadget passed to the popup menu
handler will be the GFC GGadget itself. As so, its userdata would be the
initiator's user data, and NOT the GFC struct, as is normally expected.

Thus, we cannot call GGadgetGetUserData. However, since in the GFC struct,
its GGadget comes first, effectively the pointer to its GGadget will equal
the pointer to its GFC struct.

Therefore, ensure to always pass the GFC GGadget to the popup handler and
simply cast from this to the GFC struct.