splinestroke.c: Prevent segfault/hang if BreakLine produces invalid results
I'm unsure if there is a better fix to this situation, but if the added
condition occurs, then the result will definitely be invalid, and if execution
of the function is allowed to continue, either a segfault or infinite loop
will result. This is because the following code will free first2 and second2,
so if first2 *is* second1 or second2 *is* first1, then we're essentially
use-after-freeing first1/second1.

Fixes #2832.