GDK: Filter out move messages and indicate window exists even if invisible
_GWidget_TopLevel_eh will return without processing if GDrawNativeWindowExists
returns false. This means that under the old code, for windows containing
child windows that are only properly sized by GWidgetFlowGadgets, this will
only happen when the window becomes visible and when a configure message is
received after that.

By removing the visibility check from GGDKDRawNativeWindowExists, we allow
the configure events to get send and this means that the window can be
properly sized before it becomes visible.

The second change in this commit is to filter out et_resize messages that
are only the result of moving. They are unnecessary (if any parts need
to be redrawn due to a move, an expose message will/should be sent).

Furthermore, on Windows in particular, it causes excessive CPU usage
because repeated move messages are sent as the window is moved around -
in comparison to Linux/X11 which only sends one at the end of the move,
or even OS X, which compresses the sizing events.