Fix "Unicode 2.0, full repertoire" cmap encoding.
Previously, FontForge would emit a cmap table with Platform ID = 0,
Platform Specific Encoding ID = 10. The 0/10 combination is bogus. It
should be 0/4 instead.

Both these specs:
  - https://www.microsoft.com/typography/otspec/name.htm
  - https://developer.apple.com/fonts/TrueType-Reference-Manual/RM06/Chap6cmap.html
lists 0/4 and 3/10 as valid combinations, but not 0/10. The first number
is the PID, and means Unicode (0) or Microsoft (3). The second number is
the PSID, but its meaning depends on the PID.

Fixes #2728.