Fix many of the free() calls to PyMem_Free() for the python3.6 requirement This is also based on the discussion happened in upstream issue https://github.com/fontforge/fontforge/issues/2992
This commit is a fix for the reported two problems in issue
https://github.com/fontforge/fontforge/issues/3042

The upstream documentation says
PyArg_ParseTuple() will allocate a buffer of the needed size, copy the
encoded data into this buffer and adjust *buffer to reference the newly
allocated storage. The caller is responsible for calling PyMem_Free()
to free the allocated buffer after use.