clarify python-dev requirement and force a halt if user wants it
Fix a long standing bug where configure does not differentiate
between python (the interpreter), and python-dev which is required
for adding scripting and extensions. This clarifies that configure
will check and quietly continue if no python is requested.
If user requests python scripting during configure, but there is
no python dev package, then configure will HALT with an error.