Clarify Xwindows OR GDK when choosing to use FontForge with a GUI
GDK graphical backend was recently added, and we need to organize the
options as a list to show users two GUI choices.

GDK brings its' own copies of pango and cairo, but X has the choice of
including cairo, so we show cairo as a subset of X (Pango required).
Add some more info for pango (see issues #2734, #2968, where's xft?),
and do extra header and function checks for distros that forget to go
look into /usr/local.
GDK is of better interest on the Mac, so issue #3077 not as important.

Put collab in the GUI section of the list since it is not used in the
command line version of FontForge.