Greatly improve CharView performance, see #3510
This probably shouldn't be merged, but for some reason greatly increases
the performance of the character view window. I'm mostly opening this to
try to start a discussion on why it's so slow in the first place.

"Rough" looking glyphs are quite common in fonts, so there's no reason
FontForge shouldn't be able to manipulate them.