Fix crash opening generate dialog for bitmap fonts
Opening generate font dialog for a bitmap font currently crashes when
FontForge is build without WOFF2 support. This is because existing code
assumes the last element in savefont_extensions array equals ff_none and
depends on it, but when built without WOFF2 support that array will be
one element short and bad things happen.

Fixed by making the calling code more robust, and also removing the #ifdefs
from the array; it should have the same number of elements regardless of
the build options (yay for having more optional features).