Add "UnicodeStrict" option
In many places in the code, George has put his own opinions on the
appearance of Unicode characters above the Unicode standard itself. He
may have done so rightly in some cases, but the problem is that he was
wrong in others (at least under the current version of the Unicode
Standard).

This PR adds a new option, "UnicodeStrict", which users to use the
standard above George's opinions when building characters via "Build
Accented Glyph"—there are likely more places it could be added, but this
is a start.

Close fontforge/fontforge#3615