Add `WriteModificationTime` option
By default, FontForge writes a `ModificationTime` into every SFD file.
This can disturb merging when using a version control system to manage
your font versions. Merge conflicts are inevitable with this line and
very annoying.

This PR adds an option to remove the line. I tested it, and FontForge
opens SFD files without the line without complaint or line in the
Warnings dialog.

If this PR is merged, the default will be to write the line, as is the
current action.