Coverity: several reports
Coverity indicates this (eg: cid#1083113)
1530  toolsfont = GDrawInstanciateFont(NULL,&rq);
8. noescape: Resource toolsfont is not freed or pointed-to in GResourceFindFont. [show details]
CID 1083113 (#1 of 1): Resource leak (RESOURCE_LEAK)9. overwrite_var: Overwriting toolsfont in toolsfont = GResourceFindFont("ToolsPalette.Font", toolsfont) leaks the storage that toolsfont points to.
1531  toolsfont = GResourceFindFont("ToolsPalette.Font",toolsfont);

Let's investigate:
GDrawInstanciateFont returns an object "toolfont" (which is the result of an allocation but can be NULL).

Then GResourceFindFont is called and one of its argument is "toolsfonts";
Here's its definitione:
FontInstance *GResourceFindFont(char *resourcename,FontInstance *deffont) {
    char *val = GResourceFindString(resourcename);
    if ( val==NULL )
return( deffont );
return( GResource_font_cvt(val,deffont));
}

so if "ressourcename" is found, val is not null and we call "GResource_font_cvt"
this function  (in ggadgets.c) has this declaration:
void *GResource_font_cvt(char *val, void *def)

if we focus on "def" var, we've got in the definition of this function this:
    if ( def!=NULL )
	GDrawDecomposeFont((FontInstance *)def, &rq);
    if ( fi==NULL )
return( def );
return( (void *) fi );

So if "fi" is not NULL, we don't return "def" (which corresponds to "toolsfont") but "fi"
and "def" is leaked since the return of the function will be assigned to "toolfonts"