Fix some crashes/weird behavior in gtextfield
(Please carefully review this before merging.)

This closes fontforge/fontforge#3621 and fixes fontforge/fontforge#3620,
which are both weird crashes in the "Execute Script" dialog.

It's extremely pointless to print "About to crash" before a crash, so I
totally removed that. Furthermore, despite the message, I'd never
experienced a crash after that was printed. The message goes back to
2001, in commit 78bdbeb3ef349c35e3b8f64b0689a4574391d686...I think it
can be easily seen to be outdated.

As far as the crash goes, I *believe* this totally fixes it while
retaining all previous functionality.