Add user decompositions
This closes fontforge/fontforge#3615.

This supersedes fontforge/fontforge#3617.

This allows users to set their own decomposition for each buildable
glyph.

A new line is added for each char in the SFD, `Decomposition`. It has
the same format as the `Comment` line.

Through the Python interface, users can set the decompositon using the
`glyph.user_decomp` attribute. This has been added to `python.html`.

`SFGetAlternate` now respects user decompositions if a `SplineChar` is
passed, as happens when characters are built.

Besides all the changes to "Glyph Info &rarr; Components", I unindented
"Components", so it is no longer under "Ligatures" in the menu, since
with this patch that no longer makes sense.