Make comb. chars always superior to spacing chars
Combining characters are superior in some cases to spacing characters
but not others. For example, combining macron (U+304) is superior to
macron (U+AF), yet combining dieresis (U+308) is inferior to dieresis
(U+A8).

This situation is untenable. It makes it impossible, while designing a
font, to know which character will be used when FontForge builds an
accented glyph without digging through the source code. Furthermore, the
logical case, to use the combining character first, is rarely done.

Apparently this is due to some restriction of Type1 fonts according to
George in 2001. I fixed his reported restriction in this same commit by
just dissolving the reference if the character referred to is between
0x300 and 0x345.

Some, such as @frank-trampe, might say that dissolving the reference
“breaks” something. But nothing is broken. No function of a Type 1 font
changes if a reference is used or not. Type 1 fonts never were a first
class format for storing font data—and certainly haven't become one—
that's what SFD is for. The only discernible difference therefore is a
small increase in filesize.

I confirmed:
* generated PFA (PostScript type 1) fonts still work; FontForge shows
the dissolved references as expected when a problem character is
referred to.
* if you want to use `dieresis` you can tell FontForge to in “Glyph Info
→ Components” and it will work.
* that this closes #3708, the issue is solved.

I encourage my fellow members to support this change. As an author of
five and counting open source fonts made only using FontForge and
FontTools, I can unequivocally say that sacrificing the file size of a
subset of Type 1 fonts is worth the benefit this change provides.