Fix reading glyph names from post table
Fixes https://github.com/fontforge/fontforge/issues/3740

FontForge seems to assume that glyphnames array is equal to the number
of used name indices, but in NotoSerifTibetan-Regular.ttf this is not
the case. The font has many unused name strings and the used ones start
at an index higher than the number of font glyphs, so FontForge ends
assigning the names to the glyph id 0 over and over again. Fix this by
actually reading all names in the array tell the end of the table which
seems to be what other tools do.