Make zoom & view offset tab-based (CharView) (#3772)
* Make zoom & view offset tab-based (CharView)

Close fontforge/fontforge# #3771. See the video there, and compare it to
this commit, to see why this change is needed.

* Remove xoff, yoff and scale from CharView

This revealed several places those fields were used outside of charview.c, so I
patched those.

* Make it work when FreeType debugger is enabled

I don't typically compile with --with-freetype-source so I missed this;
sorry about that!

* Fix opening of bitmap views

I accidentally broke the ability to open bitmap views by forgetting to
create a GTabSet for the `cvtemp` object. This fixes that.

* Free my CharView (@jtanx)

* Fix «Window» menu crash after closing BitmapView

* Adopt fix to BitmapView issue proposed by @skef

I only added in some GGadget trickery so I wouldn't have to change
GTabSetGetSel or CVDrawSplineSetSpecialized!

* Let's use CVGetActiveTab—make this alot less hacky

Thanks @jtanx

* Oops, fix CVDrawRubberRect

* Remove `CharViewTab cvtabtemp` per @jtanx

* Consistency—CVGetActiveTab everywhere! per @jtanx