Fix wordlist error w/glyph name matching ^uni & \.
This fixes a bug wherein the word list input in CharView was unable to
understand input like [uni0061.alt]a, fudging it into
/uni0061/periodalta. This is because the parser was simply giving up
after trying to find "uni0061.alta", instead of searching for
"uni0061.alt".

The meat and potatoes of this commit is the added for loop. The rest is
just whitespace fixes to make this part of the code actually readable.