Fold libuninameslist into FontForge
Clean this mess up by including the two relevant library files into the
source tree (and the builder tool, but not compiled, this C mess should
be rewritten as a Python script).

The French translation is not used by FontForge, so not included.

Support for building with external Unicode names library is dropped,
this obfuscated mess of C macros is an overkill for such a tiny FontForge
functionality.

Probably more cleanup can be done, but that is for another day.