Stop crash under tiling WMs
This closes #3920.

I ran through this with `rr` and can see this change to be the correct
one. This does the same thing as the patch I offered in #3920, except is
much more readable.

When the window is dying, it has no widget_data to free; the free was
already done by `_GWidget_Container_eh`. The order of events is
different on tiling window managers