Fix crash in SFDTrimUndoOldToNew.
The innermost 'while' loop advanced 'oglyph' with
'SFDMoveToNextStartChar' but did not check the return value for 0 (to
check for the last entry). As a result, 'SCFindByGlyphName' could crash
in strcmp. This changes 'SCFindByGlyphName' not to crash in this
condition and also checks for oglyph being 0 where needed. This fixes
crashes while editing font lookups.