Fix non-terminating AddExtrema loop (Expand Stroke) (#4058)
ForceEndPointExtrema is supposed to return -1 if it can't find anything to do. In some cases the control point position will converge on the on-curve point position, but SPAdjustControl() (in virtue of SplineRefigureFixup(), in virtue of IntersectLinesClip()) will move the cp slightly away from to. This creates a non-terminating loop where ForceEndPointExtrema() thinks it's moving the cp but isn't.

The solution in this PR is to check to see if it's moved and adjust the return value accordingly.