Always allow opening «Glyph Info…» in FontView
This is another one of those issues “I’d gotten so used to I forgot they
were actually bugs.”

In master it is not possible to open this dialog for glyphs which don't
exist…via the keyboard, or Element menu. But you can just right-click
the glyph, and receive the dialog.

Which I do…often. The restriction seems to have no purpose, and is very
annoying, so I've removed it.