Clarify OpenCharsInNewWindow
I believe this was originally a bug, but due to how long it has endured,
it has now become a feature.

Consider:

    * if OpenCharsInNewWindow be false, and;
    * if a char be double-clicked,

it will open a new tab in an existing window. But if all the same is
true, except that the user pressed Return instead, a new window is
opened, despite the setting of OpenCharsInNewWindow.

So, we need a way to open new CharView tabs from the keyboard. I decided
Shift+Enter is just as good as any other way.