Fix spelling mistake in 3D projection tool
A plain is a flat, sweeping landmass that does not change much in elevation. It's a geographical term. See  also https://en.wikipedia.org/wiki/Plain
A plane is a flat, two-dimensional surface in mathematics. This is more appropriate if you're not talking about landscapes or horses or whatever. See also https://en.wikipedia.org/wiki/Plane_(geometry)