Make it possible to ask user for multiple choices (Python) (#4143)
Multiple of which can be selected.

This is needed for #703, which I'm implementing as a Python script.

This is the first of two new needed APIs. The second will add a new
function that works on `fontforge.glyph` objects, similar to `build`,
but will instead take a combining class type and place an anchor.