Add fontforge.scriptFromUnicode(n) (Python)
Some scripts need to add OpenType tables without knowing which OpenType
script tag is needed for a particular glyph or set of glyphs. It's true
that there's always `DFLT`, but that doesn't work in FontForge's UI by
default, and presumably it doesn't work with some shapers either.

This function is in the theme of the other UnicodeNames* functions, so
it shouldn't be a problem to include it.