fontview.c: ensure glyph centering is done wrt bdfc->width and not base.width
This fixes a discrepancy in viewing compressed vs uncompressed bitmaps (i.e.
bitmaps run through BCCompressBitmap), where compressed in this sense means no
padding (unused rows or columns around the outside).

For the case of a compressed bitmap, the bitmap size may be smaller than the
actual width, which leads to it being centered differently in the fontview when
base.width is used.

So instead, use bdfc->width to ensure consistency between compressed/uncompressed
bitmaps. For a compressed bitmap, bdfc->xmin should represent the 'true' offset
to the first used column, so to achieve the same padding that would have been
present for an uncompressed bitmap, subtract off 2*bdfc->xmin.

The 2x factor comes from this:

The x offset is calculated from:
(fv->cbw-1-xwidth)/2

For a compressed bitmap where we know the true x offset, this would
theoretically be equivalent:
(fv->cbw-1-bdfc->width)/2 + bdfc->xmin

But because of integer rounding, the xmin offset needs to go beforehand,
hence the 2x scaling up and negation.

See also: #4262