Fix pfadecrypt bugs (#5074)
* Fix typo in decodestr()

The typo resulted in the generation of complete gibberish when decrypting glyph charstrings.

* Fix decryption of binary eexec ciphertext

A font's binary eexec ciphertext is followed by 512 clear text '0' characters which may be interspersed with white space characters.  Typically, there is a newline character preceding the first '0'.  The program was treating that newline as a final character of the ciphertext rather than as part of the white space.