Add option to ignore non-existent glyphs in substitution rules of OpenType feature file
When reference to non-existent glyphs are present in an OpenType feature file, merging the feature fails.
Add an option to skip the erroring out in merging feature file. This helps with feature files
of Indic scripts with many hundreds of ligatures; some of the glyphs may not be present in all fonts,
yet the feature file can be reused.

TODO: handle the case when an entire lookup list is invalid.